alibaba-backend/auth-service/requirement.txt
fastapi
uvicorn[standard]
psycopg2-binary
psycopg2
fastapi-login
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart